Refactored to widgets Enabled chat list pane width resize Fixed adaptive resize on resize window
