+++ homebase-common/src/commonMain/composeResources/values-da/strings.xml
@@ -1,7 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
+    <!-- Menu options -->
     <string name="settings">Indstillinger</string>
-    <string name="chat_select_a_conversation">Vælg en samtale</string>
+
+    <!-- Language options -->
     <string name="language">Sprog</string>
     <string name="language_system">Systemstandard</string>
     <string name="language_english_us">Engelsk (USA)</string>
@@ -22,4 +24,10 @@
     <string name="day_friday">Fredag</string>
     <string name="day_saturday">Lørdag</string>
     <string name="day_sunday">Søndag</string>
+
+    <!-- Chat list -->
+    <string name="chat_select_a_conversation">Vælg en samtale</string>
+    <string name="chat_select_a_conversation_subtitle">Vælg en samtale fra listen for at se beskederne</string>
+    <string name="chat_new_conversation">Ny samtale</string>
+    <string name="chat_search_placeholder">Søg</string>
 </resources>
