+++ homebase-common/src/commonMain/composeResources/values/strings.xml
@@ -1,8 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="app_name">Homebase Chat</string>
+
+    <!-- Menu options -->
     <string name="settings">Settings</string>
-    <string name="chat_select_a_conversation">Select a conversation</string>
+
+    <!-- Language options -->
     <string name="language">Language</string>
     <string name="language_system">System Default</string>
     <string name="language_english_us">English (USA)</string>
@@ -23,4 +26,10 @@
     <string name="day_friday">Friday</string>
     <string name="day_saturday">Saturday</string>
     <string name="day_sunday">Sunday</string>
+
+    <!-- Chat list -->
+    <string name="chat_select_a_conversation">Select a conversation</string>
+    <string name="chat_select_a_conversation_subtitle">Choose a conversation from the list to view messages</string>
+    <string name="chat_new_conversation">New conversation</string>
+    <string name="chat_search_placeholder">Search</string>
 </resources>
