+++ homebase-core/src/commonMain/kotlin/id/homebase/core/App.kt
@@ -53,5 +53,4 @@ fun App(
             onNavHostReady(navController)
         }
     }
-
 }
